refactor: move tests cards to src/components/cards
