+++ dashboard/src/components/Cards/ErrorCountList.tsx
@@ -1,0 +1,38 @@
+import { memo } from 'react';
+
+import type { IBaseCard } from '@/components/Cards/BaseCard';
+import BaseCard from '@/components/Cards/BaseCard';
+import { DumbListingContent } from '@/components/ListingContent/ListingContent';
+import ListingItem from '@/components/ListingItem/ListingItem';
+import type { TTreeTestsData } from '@/types/tree/TreeDetails';
+
+interface IErrorCountList extends Pick<TTreeTestsData, 'errorMessageCounts'> {
+  title: IBaseCard['title'];
+}
+
+const ErrorCountList = ({
+  errorMessageCounts,
+  title,
+}: IErrorCountList): JSX.Element => {
+  return (
+    <BaseCard
+      title={title}
+      content={
+        <DumbListingContent>
+          {Object.keys(errorMessageCounts).map(errorMessage => {
+            const currentErrorMessageCount = errorMessageCounts[errorMessage];
+            return (
+              <ListingItem
+                key={errorMessage}
+                text={errorMessage}
+                errors={currentErrorMessageCount}
+              />
+            );
+          })}
+        </DumbListingContent>
+      }
+    />
+  );
+};
+
+export default memo(ErrorCountList);
